+++ src/components/YearChart.vue
@@ -124,6 +124,7 @@ export default {
 .year-chart-svg {
     width: 100%;
     height: auto;
+    max-height: 220px;
 }
 
 .bar {